requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.28.0
ultralytics
